:recycle: Refactor bool serialization to the wasm module
