+++ render-wasm/src/wasm/paths.rs
@@ -1,15 +1,14 @@
 #![allow(unused_mut, unused_variables)]
-use indexmap::IndexSet;
 use macros::ToJs;
 use mem::SerializableResult;
-use uuid::Uuid;
+use std::mem::size_of;
 
-use crate::math::bools;
-use crate::shapes::{BoolType, Path, Segment, ToPath};
-use crate::uuid;
-use crate::{mem, with_current_shape, with_current_shape_mut, with_state, STATE};
+use crate::shapes::{Path, Segment, ToPath};
+use crate::{mem, with_current_shape, with_current_shape_mut, STATE};
 
 const RAW_SEGMENT_DATA_SIZE: usize = size_of::<RawSegmentData>();
+
+pub mod bools;
 
 #[repr(C, u16, align(4))]
 #[derive(Debug, PartialEq, Clone, Copy, ToJs)]
@@ -180,37 +179,6 @@ pub extern "C" fn current_to_path() -> *mut u8 {
             .collect();
     });
 
-    mem::write_vec(result)
-}
-
-#[no_mangle]
-pub extern "C" fn calculate_bool(raw_bool_type: u8) -> *mut u8 {
-    let bytes = mem::bytes_or_empty();
-
-    let entries: IndexSet<Uuid> = bytes
-        .chunks(size_of::<<Uuid as SerializableResult>::BytesType>())
-        .map(|data| Uuid::from_bytes(data.try_into().unwrap()))
-        .collect();
-
-    mem::free_bytes();
-
-    let bool_type = BoolType::from(raw_bool_type);
-    let result;
-    with_state!(state, {
-        let path = bools::bool_from_shapes(
-            bool_type,
-            &entries,
-            &state.shapes,
-            &state.modifiers,
-            &state.structure,
-        );
-        result = path
-            .segments()
-            .iter()
-            .copied()
-            .map(RawSegmentData::from_segment)
-            .collect();
-    });
     mem::write_vec(result)
 }
